+++ client/package.json
@@ -13,7 +13,9 @@
     "axios": "^1.4.0",
     "bootstrap": "^5.3.1",
     "core-js": "^3.8.3",
+    "pinia": "^2.1.6",
     "vue": "^3.2.13",
+    "vue-router": "^4.2.4",
     "vue3-apexcharts": "^1.4.4"
   },
   "devDependencies": {
