+++ client/src/main.js
@@ -1,9 +1,13 @@
 import { createApp } from 'vue'
+import { createPinia } from 'pinia';
 import VueApexCharts from 'vue3-apexcharts';
-import "bootstrap/dist/css/bootstrap.min.css"
-import "bootstrap"
+import { router } from './router';
 import App from './App.vue'
 
 const app = createApp(App);
+
 app.use(VueApexCharts);
+app.use(createPinia());
+app.use(router);
+
 app.mount('#app');
